workaround for spending account bug - EX-1053
When a payment with a type spending account is deleted the amount on that spending account gets added to the spending account. This workaround first patches the payment to type purchase order and then deletes it. Revert this back to just deleting the payment once the api bug has been resolved

Resolve: OC-2084