F51-185 Refactor payment directives into one
There is now one payment directive for handling both multiple
and single payments of any given type. The directive will default
to the checkout config constants if the attributes are not used.

Attributes:
- multiple: boolean default, false
- methods: array, default, all
- order: current order (required)

This directive now lives in the common/directives folder
and the ocPayments service has been moved to common/services.